refactor: Change from StartupActivity to ProjectActivity
